// webapps/admin/generated/http-clients/backend/models/participants-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shula API
 * API documentation for the Shula application
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ParticipantsResponseDto
 */
export interface ParticipantsResponseDto {
    /**
     * Unique identifier
     * @type {string}
     * @memberof ParticipantsResponseDto
     */
    'id': string;
    /**
     * Participant phone number
     * @type {string}
     * @memberof ParticipantsResponseDto
     */
    'phoneNumber': string;
    /**
     * Participant name
     * @type {string}
     * @memberof ParticipantsResponseDto
     */
    'name': string;
    /**
     * Participant status
     * @type {string}
     * @memberof ParticipantsResponseDto
     */
    'status': string;
    /**
     * Date when participant joined
     * @type {string}
     * @memberof ParticipantsResponseDto
     */
    'joinedAt': string;
}